Add msg codec (#7)
* init

* add encode & decode to msg

* add const values

* add framed udp socket

* remove comments

* remove commetns

* Add W

+++ src/common/connection.rs
@@ -1,5 +1,5 @@
 #[derive(Clone)]
 pub struct Conection {
-    pub id: u32,
+    pub id: u128,
     pub ip: String,
 }
